requirements.txt
absl-py==2.1.0
alembic==1.14.0
astunparse==1.6.3
bcrypt==4.2.1
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
coverage==7.6.10
cryptography==44.0.0
decorator==5.1.1
Deprecated==1.2.15
fabric==3.2.2
filelock==3.16.1
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Migrate==4.0.7
Flask-MySQL==1.6.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
flatbuffers==24.12.23
fsspec==2024.12.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.68.1
h5py==3.12.1
idna==3.10
importlib_metadata==8.5.0
invoke==2.2.0
itsdangerous==2.2.0
Jinja2==3.1.5
keras==3.7.0
libclang==18.1.1
Mako==1.3.8
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
namex==0.0.8
networkx==3.2.1
numpy==2.0.2
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
paramiko==3.5.0
protobuf==5.29.2
pycparser==2.22
Pygments==2.18.0
PyMySQL==1.1.1
PyNaCl==1.5.0
requests==2.32.3
rich==13.9.4
six==1.17.0
SQLAlchemy==2.0.36
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
torch==2.5.1
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.0
WTForms==3.2.1
zipp==3.21.0
